Split bfffsd/fs.rs's unit tests into a separate file.
